changed state type to cell

+++ life_backend.hs
@@ -6,9 +6,9 @@ import System.Random
 -- some solutions could be: - have a set width and height of the board - not a huge fan since the board should be infinite in the game of life
 --                          - we trim it when an area is all "dead" - I like this one more and we can probably implement it later rather than sooner
 
-type Position = (Int,Int)
+type Position = (Float,Float)
 
-data CellState = Alive | Dead 
+data CellState = Alive | Dead
     deriving (Eq, Show)
 
 data Cell = Cell CellState Position
@@ -35,7 +35,7 @@ getCellPositions cells = [position | (Cell state position) <- cells]
 
 -- (-1,-1) (0, -1) (1, -1)
 -- (-1, 0)         (1 , 0) <- relative adjacent coordinates
--- (-1, 1) (0 , 1) (1 , 1) 
+-- (-1, 1) (0 , 1) (1 , 1)
 
 -- get the world coordinates of adjacent cells to a given position
 getAdjacents :: Position -> [Position]
@@ -43,15 +43,15 @@ getAdjacents (x,y) = map (\ (i,j) -> (i+x, j+y)) adjacent
     where adjacent = [(-1,-1), (0, -1), (1, -1), (-1, 0), (1 , 0), (-1, 1), (0 , 1), (1 , 1)]
 
 -- tests
--- getAdjacents (0,0) 
+-- getAdjacents (0,0)
 -- getAdjacents (3,4)
 
 
 -- given a Cell and the Board it resides on, return a list of it's neighbouring Cells
 neighbours :: Cell -> Board -> [Cell]
-neighbours (Cell _ position) board = 
+neighbours (Cell _ position) board =
     currAdjacents ++ otherAdjacents
-    where 
+    where
         adjacentPositions = getAdjacents position
         currAdjacents = [(Cell state position) | (Cell state position) <- board,  position `elem` adjacentPositions] -- Already existing cells on board
         otherAdjacents = [(Cell Dead position) | position <- adjacentPositions, not (position `elem` (getCellPositions currAdjacents))] -- Cells currently not on the board
@@ -68,18 +68,18 @@ isAlive (Cell state _)
     | state == Alive = True
     | otherwise = False
 
--- counts the number of occurances in the list that matches parameter p 
+-- counts the number of occurances in the list that matches parameter p
 count :: Num p => (t -> Bool) -> [t] -> p
 count p [] = 0
 count p (h:t) = if p h then 1 + count p t else count p t
-    
+
 -- !!TODO!!: given a cell and it's neighbours (and a probability?), return the cell in the next generation
 -- A cell with 2 or 3 live neighbours survives
 -- A dead cell with 3 live neighbours becomes live
 -- All other cells become dead. Dead cells stay dead
 --nextCellGeneration :: Cell -> [Cell] -> Cell
 --nextCellGeneration (Cell state position) nb = if state == Alive then stillAlive (Cell Alive position) nb else stillDead (Cell Dead position) nb
---   where 
+--   where
 --       stillAlive c nb = if count (==True) (map isAlive nb) `elem` [2,3] then (Cell Alive position) else (Cell Dead position)
 --       stillDead c nb = if count (==True) (map isAlive nb) == 3 then (Cell Alive position) else (Cell Dead position)
 
@@ -91,7 +91,7 @@ count p (h:t) = if p h then 1 + count p t else count p t
 nextCellGenP :: Cell -> [Cell] -> Int -> Cell
 nextCellGenP (Cell state position) nb prob = if state == Alive then stillAlive (Cell Alive position) nb prob else stillDead (Cell Dead position) nb prob
 
--- Determines if a living cell remains alive 
+-- Determines if a living cell remains alive
 -- A cell with 2 or 3 living neighbours remains alive, otherwise the cell dies with the given probability
 stillAlive :: Cell -> [Cell] -> Int -> Cell
 stillAlive (Cell state position) nb prob = if count (==True) (map isAlive nb) `elem` [2,3] then (Cell Alive position) else switch (Cell Alive position) prob
@@ -106,11 +106,11 @@ switch :: Cell -> Int -> Cell
 switch (Cell state position) a = if (state == Alive && a == 1) || (state == Dead && a == 0) then (Cell Dead position) else (Cell Alive position)
 
 
--- Given a probability, p (1-100), generates a list with p 1's and (100-p) 0's and chooses a value at random 
+-- Given a probability, p (1-100), generates a list with p 1's and (100-p) 0's and chooses a value at random
 --pick :: Num b => Int -> IO b
 --pick p = (\index -> (replicate p 1 ++ (replicate (100-p) 0)) !! index) <$> randomRIO (0,99)
 pick :: Num a => Int -> Int -> a
-pick p index = (replicate p 1 ++ (replicate (100-p) 0)) !! index 
+pick p index = (replicate p 1 ++ (replicate (100-p) 0)) !! index
 
 --tests
 --c1 = Cell Alive (0,1)
@@ -123,7 +123,7 @@ pick p index = (replicate p 1 ++ (replicate (100-p) 0)) !! index
 --nb = neighbours c3 board
 --neigh = neighbours c6 board
 
--- Generates next board 
+-- Generates next board
 nextBoardGen:: Board -> Int -> [Int] -> Board
 nextBoardGen board prob randList = [nextCellGenP cell (neighbours cell board) (pick prob (randList !! i)) | (i, cell) <- (zip [0..] board)]
 
@@ -144,7 +144,7 @@ allDead board = all (==Dead) [state | (Cell state position) <- board]
 -- if all dead then return EndOfGame
 -- if not then calculate the next board and return it as part of ContinueGame
 gameOfLife :: Board -> Int -> IO Result
-gameOfLife board p = 
+gameOfLife board p =
     do
         rg <- newStdGen
         let randList = randomRs (0,99) rg
@@ -168,7 +168,7 @@ gameOfLife board p =
 
 -- play board 0 = return board
 
--- play board n = 
+-- play board n =
 --     do
 --         let result = gameOfLife board
 --         if (result == EndOfGame)
@@ -176,4 +176,3 @@ gameOfLife board p =
 --                 return newBoard
 --             else
 --                 play newBoard (n-1)
-
